Allowing in-memory alternative to mongodb
