+++ electron/package.json
@@ -11,7 +11,6 @@
     "cookie-parser": "^1.4.5",
     "express": "^4.17.1",
     "express-async-handler": "^1.1.4",
-    "mongodb": "^3.5.5",
     "node-cache": "^5.1.0",
     "node-hue-api": "^4.0.5",
     "nodemailer": "^6.4.6"
