+++ server/db/memory/connection.ts
@@ -1,0 +1,4 @@
+
+export async function initClient () {
+  // noop
+}
