+++ server/db/memory/oauthState.ts
@@ -1,0 +1,10 @@
+
+import { GenerateOauthStateFn, VerifyOauthStateFn } from '../types/OAuthState'
+
+export const generateOauthState: GenerateOauthStateFn = async function () {
+  return ''
+}
+
+export const verifyOauthState: VerifyOauthStateFn = async function () {
+  return false
+}
