+++ server/db/memory/remoteCredentials.ts
@@ -1,0 +1,14 @@
+
+import { SaveRemoteCredentialsFn, GetRemoteCredentialsFn, DeleteRemoteCredentialsFn } from '../types/RemoteCredentials'
+
+export const saveRemoteCredentials: SaveRemoteCredentialsFn = async function () {
+  // noop
+}
+
+export const getRemoteCredentials: GetRemoteCredentialsFn = async function () {
+  return null
+}
+
+export const deleteRemoteCredentials: DeleteRemoteCredentialsFn = async function () {
+  return false
+}
